Investigating the CAM
Through this discovery, you will investigate the contents and properties of the CAM table of a switch. You can display the CAM table to verify the information that the switch has learned. The table will also tell you out of which port the address was learned, along with the VLAN information.